doc-rec/requirements.txt
pytesseract==0.3.9
Pillow==9.1.1
pydantic==1.9.1
requests==2.27.1
environs==9.5.0
loguru==0.6.0
pdf2image==1.16.0
Flask==2.1.2
Flask-Pydantic==0.9.0